+++ src/build_info.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+
+namespace bpftrace {
+
+class BuildInfo
+{
+public:
+  static std::string report();
+};
+
+} // namespace bpftrace